clean change password & move to own component & now used server feedback and better rules & fix the layout
